+++ contracts/cosmwasm-vm/cw-ibc-core/src/ics03_connection/mod.rs
@@ -10,8 +10,10 @@ use crate::ics03_connection::event::create_open_init_event;
 use crate::ics03_connection::event::create_open_try_event;
 use crate::types::{ClientId, ConnectionId};
 use crate::types::{OpenAckResponse, OpenTryResponse};
-use crate::types::{VerifyClientConsesnusState, VerifyClientFullState, VerifyConnectionState};
 use crate::ContractError;
+use common::client_msg::{
+    VerifyClientConsesnusState, VerifyClientFullState, VerifyConnectionState,
+};
 use cosmwasm_std::DepsMut;
 use cosmwasm_std::Event;
 use cosmwasm_std::Response;
